requirements.txt
Flask==3.0.0
gunicorn==21.2.0
mysql-connector-python==8.2.0
